requirements/base.txt
djangorestframework==3.14.0
Django==4.0.8
django-cors-headers==3.13.0
django-environ==0.9.0
drf-yasg[validation]
typing-extensions
Pillow==9.2.0

redis==4.3.4
psycopg2-binary==2.9.4

celery==5.2.7

pyspellchecker==0.7.0
pymorphy2
